move Browse, refactor
