+++ examples/Web/web/src/App.css
@@ -125,36 +125,3 @@ body {
     width: 1200px !important;
     margin: auto !important;
 }
-
-.browse-results {
-    width: 1200px !important;
-    margin: auto !important;
-}
-
-.browse-folderlist {
-    width: 100% !important;
-    padding-left: 0px !important;
-    white-space: pre;
-    overflow: auto !important;
-    max-height: calc(50vh - 100px);
-}
-
-.browse-folderlist-list {
-    padding-left: 15px !important;
-    padding-top: 15px !important;
-}
-
-.browse-filelist {
-    width: 100% !important;
-    padding-right: unset !important;
-}
-
-.ui.list .list, ol.ui.list ol, ul.ui.list ul {
-    padding-left: 0px !important;
-    padding-top: 3px !important;
-    padding-bottom: 0px !important;
-}
-
-.ui.list .list>.item>.icon+.content, .ui.list .list>.item>.image+.content, .ui.list>.item>.icon+.content, .ui.list>.item>.image+.content {
-    padding-left: 0px !important;
-}
